refactor routing structure for authentication and dashboard modules; consolidate routing files
