+++ src/app/modules/auth/auth.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { AuthLayoutComponent } from '../../presentation/layouts/auth-layout/auth-layout.component';
+import { AuthRoutingModule } from '../routes/auth-routing.module';
 
-import { AuthRoutingModule } from './auth-routing.module';
-import { AuthLayoutComponent } from '../../../presentation/layouts/auth-layout/auth-layout.component';
+
 
 
 @NgModule({
