+++ src/app/modules/dashboard/dashboard.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { DashboardLayoutComponent } from '../../presentation/layouts/dashboard-layout/dashboard-layout.component';
+import { DashboardRoutingModule } from '../routes/dashboard-routing.module';
 
-import { DashboardRoutingModule } from './dashboard-routing.module';
-import { DashboardLayoutComponent } from '../../../presentation/layouts/dashboard-layout/dashboard-layout.component';
+
 
 
 @NgModule({
