+++ src/app/modules/routes/auth-routing.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AuthLayoutComponent } from '../../../presentation/layouts/auth-layout/auth-layout.component';
+import { AuthLayoutComponent } from '../../presentation/layouts/auth-layout/auth-layout.component';
+
 
 const routes: Routes = [
   { path: '',
