+++ src/app/modules/routes/dashboard-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { DashboardLayoutComponent } from '../../presentation/layouts/dashboard-layout/dashboard-layout.component';
 
 
-import { DashboardLayoutComponent } from '../../../presentation/layouts/dashboard-layout/dashboard-layout.component';
 
 const routes: Routes = [
   { path: '',
